rm upgrade handler and move migration helpers from app/upgrades to x/staking
