--- app/upgrades/v11/constants.go
@@ -1,15 +1,0 @@
-package v11
-
-import (
-	"github.com/cosmos/gaia/v11/app/upgrades"
-)
-
-const (
-	// UpgradeName defines the on-chain upgrade name.
-	UpgradeName = "v11"
-)
-
-var Upgrade = upgrades.Upgrade{
-	UpgradeName:          UpgradeName,
-	CreateUpgradeHandler: CreateUpgradeHandler,
-}
